+++ cardano-node/src/Cardano/Node/Startup.hs
@@ -87,6 +87,10 @@ data StartupTrace blk =
   --
   | NetworkConfigUpdateError Text
 
+  -- | Legacy topology file format is used.
+  --
+  | NetworkConfigLegacy
+
   -- | Log peer-to-peer network configuration, either on startup or when its
   -- updated.
   --
@@ -96,11 +100,6 @@ data StartupTrace blk =
 
   -- | Warn when 'EnableP2P' is set.
   | P2PWarning
-
-  -- | Warn that peer-to-peer requires
-  -- 'TestEnableDevelopmentNetworkProtocols' to be set.
-  --
-  | P2PWarningDevelopementNetworkProtocols
 
   -- | Warn when 'TestEnableDevelopmentNetworkProtocols' is set.
   --