+++ cardano-node/src/Cardano/Node/Tracing/Tracers/Consensus.hs
@@ -67,6 +67,10 @@ module Cardano.Node.Tracing.Tracers.Consensus
   , severityKeepAliveClient
   , docKeepAliveClient
 
+  , namesConsensusStartupError
+  , severityConsensusStartupError
+  , docConsensusStartupError
+
   ) where
 
 
@@ -89,6 +93,7 @@ import           Cardano.Node.Tracing.Era.Shelley ()
 import           Cardano.Node.Tracing.Formatting ()
 import           Cardano.Node.Tracing.Render
 import           Cardano.Node.Tracing.Tracers.StartLeadershipCheck
+import           Cardano.Node.Tracing.Tracers.ConsensusStartupException
 import           Cardano.Prelude hiding (All, Show, show)
 
 import           Cardano.Protocol.TPraos.OCert (KESPeriod (..))
@@ -124,7 +129,6 @@ import           Ouroboros.Consensus.Node.Run (SerialiseNodeToNodeConstraints, e
 import           Ouroboros.Consensus.Node.Tracers
 import qualified Ouroboros.Consensus.Protocol.Ledger.HotKey as HotKey
 import           Ouroboros.Consensus.Util.Enclose
-
 
 
 instance LogFormatting a => LogFormatting (TraceLabelCreds a) where
@@ -557,11 +561,12 @@ instance (HasHeader header, ConvertRawHash header) =>
     mconcat [ "kind" .= String "AddedFetchRequest" ]
   forMachine _dtal BlockFetch.AcknowledgedFetchRequest {} =
     mconcat [ "kind" .= String "AcknowledgedFetchRequest" ]
-  forMachine _dtal (BlockFetch.SendFetchRequest af) =
+  forMachine _dtal (BlockFetch.SendFetchRequest af gsv) =
     mconcat [ "kind" .= String "SendFetchRequest"
             , "head" .= String (renderChainHash
                                  (renderHeaderHash (Proxy @header))
                                  (AF.headHash af))
+            , "deltaq" .= toJSON gsv
             , "length" .= toJSON (fragmentLength af)]
    where
      -- NOTE: this ignores the Byron era with its EBB complication:
@@ -1681,6 +1686,20 @@ instance Show remotePeer => LogFormatting (TraceKeepAliveClient remotePeer) wher
 
 docKeepAliveClient :: Documented (TraceKeepAliveClient peer)
 docKeepAliveClient = Documented [
+    DocMsg
+      []
+      []
+      ""
+  ]
+
+namesConsensusStartupError :: ConsensusStartupException -> [Text]
+namesConsensusStartupError _ = []
+
+severityConsensusStartupError :: ConsensusStartupException -> SeverityS
+severityConsensusStartupError _ = Critical
+
+docConsensusStartupError :: Documented ConsensusStartupException
+docConsensusStartupError = Documented [
     DocMsg
       []
       []